Refactor validators
